+++ app/Services/RussianWordService.php
@@ -15,15 +15,13 @@ class RussianWordService {
     {
         $word = new RussianWord();
 
-        $word->word = $request->getWord();
-        $word->translate = $request->getTranslate();
+        $word->russian = $request->getRussian();
+        $word->tj = $request->getTj();
+        $word->incorrect_answers = $request->getIncorrectAnswers();
 
 
         if($word->save()){
-
-            $word->words()->sync($request->getWords());
             $word->categories()->sync($request->getCategories());
-
             return $word;
         }
 
@@ -41,12 +39,12 @@ class RussianWordService {
     {
         $word = RussianWord::find($id);
 
-        $word->word = $request->getWord();
-        $word->translate = $request->getTranslate();
+        $word->russian = $request->getRussian();
+        $word->tj = $request->getTj();
+        $word->incorrect_answers = $request->getIncorrectAnswers();
 
         if($word->update()){
 
-            $word->words()->sync($request->getWords());
             $word->categories()->sync($request->getCategories());
 
             return $word;