+++ database/migrations/2021_11_14_145221_create_russian_words_table.php
@@ -15,8 +15,9 @@ class CreateRussianWordsTable extends Migration
     {
         Schema::create('russian_words', function (Blueprint $table) {
             $table->id();
-            $table->string('word');
-            $table->string('translate');
+            $table->string('russian');
+            $table->string('tj');
+            $table->string('incorrect_answers'); // json
             $table->timestamps();
         });
     }